requirements.txt
aioquic==1.2.0
allure-pytest==2.15.3
allure-python-commons==2.15.3
allure-behave==2.15.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
asttokens==3.0.1
backcall==0.2.0
beautifulsoup4==4.14.3
playwright==1.40.0
behave==1.3.3
bleach==6.3.0
blinker==1.9.0
Booktype==1.5
brotli==1.2.0
click==8.3.1
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.0.0
decorator==5.2.1
defusedxml==0.7.1
Django==6.0.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
gherkin-official==29.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
ipython==8.12.3
jsonpath-ng>=1.6.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kaitaistruct==0.11
ldap3==2.9.1
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
msgpack==1.1.2
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
pandocfilters==1.5.1
parse==1.20.2
parse_type==0.6.6
parso==0.8.5
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
publicsuffix2==2.20191221
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydivert==2.1.0
pylsqpack==0.3.23
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.11.0
pytest-bdd==8.1.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.20.2
sortedcontainers==2.4.0
soupsieve==2.8.3
sqlparse==0.5.5
sshtunnel==0.4.0
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.3
urwid==3.0.3
wcwidth==0.3.5
webencodings==0.5.1
Werkzeug==3.1.5
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
yarg==0.1.9
zstandard==0.25.0
Appium-Python-Client==2.11.1
selenium==4.16.0
